api/requirements.txt
argparse
asyncpg
fastapi>=0.100.0
python-dotenv
requests


uvicorn[standard]